# swagger/demo_management_service.yaml
openapi: 3.0.0
info:
  title: Demo Management Service API
  version: 2.0.1
  description: API for managing demos, including creation, updates, member assignments, and demo lifecycle operations

servers:
  - url: https://159.203.10.100:8501

paths:

  # Demo endpoints
  /demos/:
    get:
      tags:
        - Demo
      summary: Get all demos
      operationId: listDemos
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: Number of records to skip
          example: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
          required: false
          description: Number of items to return
          example: 100
        - in: query
          name: order_by
          schema:
            type: string
            default: "-created_at"
          required: false
          description: Field to order by. Prefix with '-' for descending order
          example: "-created_at"
        - in: query
          name: search
          schema:
            type: string
          required: false
          description: Search query string to filter results
          example: "Marketing"
        - in: query
          name: filters
          schema:
            type: array
            items:
              type: string
          required: false
          description: List of filter dicts as JSON strings
          example: '["{\"field\": \"status\", \"operator\": \"eq\", \"value\": \"created\"}", "{\"field\": \"is_active\", \"operator\": \"eq\", \"value\": true}"]'
        - in: query
          name: status
          schema:
            type: string
            enum: ['creating', 'created', 'updating', 'updated', 'deleting', 'deleted']
          required: false
          description: Filter demos by status
          example: "created"
      responses:
        '200':
          description: List of demos retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DemoModel'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/create/:
    post:
      tags:
        - Demo
      summary: Create a new demo
      operationId: createDemo
      description: | 
        **publishes** a message to the internal **`demo_creation_topic`** queue.
        Message format:
        ```json
        {
          "demo_id": "uuid"
        }
        ```
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
                  example: "Marketing Team"
              required:
                - name
      responses:
        '201':
          description: Demo created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demo created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/read/{demo_id}/:
    get:
      tags:
        - Demo
      summary: Get a demo by ID
      operationId: getDemo
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        - in: path
          name: demo_id
          required: true
          schema:
            type: string
            format: uuid
          description: Demo identifier
          example: "bd08612-1a9d-4a6e-8431-1demoid"
        
      responses:
        '200':
          description: Demo retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DemoModel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/update/{demo_id}/:
    patch:
      tags:
        - Demo
      summary: Update a demo
      operationId: updateDemo
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        - in: path
          name: demo_id
          required: true
          schema:
            type: string
            format: uuid
          description: Demo identifier
          example: "bd08612-1a9d-4a6e-8431-1demoid"
        
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
                  example: "Marketing Team Updated"
      responses:
        '200':
          description: Demo updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demo updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/delete/{demo_id}/:
    delete:
      tags:
        - Demo
      summary: Delete a demo
      operationId: deleteDemo
      description: |
        **publishes** a message to the internal **`demo_deletion_topic`** queue.
        Message format:
        ```json
        {
          "demo_id": "uuid"
        }
        ```
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        - in: path
          name: demo_id
          required: true
          schema:
            type: string
            format: uuid
          description: Demo identifier
          example: "bd08612-1a9d-4a6e-8431-1demoid"
        
      responses:
        '200':
          description: Demo deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demo deleted successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/update/status/{demo_id}/:
    patch:
      tags:
        - Demo
      summary: Update demo status and error messages
      operationId: updateDemoStatus
      description: |
        Updates the demo status and error messages.
        This endpoint allows updating only the status, error_message, and error_user_message fields.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        - in: path
          name: demo_id
          required: true
          schema:
            type: string
            format: uuid
          description: Demo identifier
          example: "bd08612-1a9d-4a6e-8431-1demoid"
        - in: query
          name: status
          required: true
          schema:
            type: string
          description: The new status for the demo
          example: "created"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: ['creating', 'created', 'updating', 'updated', 'deleting', 'deleted']
                  example: "updated"
                error_message:
                  type: string
                  nullable: true
                  description: Technical error message for debugging
                  example: "Database connection timeout after 30 seconds"
                error_user_message:
                  type: string
                  nullable: true
                  description: User-friendly error message for display
                  example: "Something went wrong. Please try again later."
      responses:
        '200':
          description: Demo status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demo status updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /demo/update/is-active/{demo_id}/:
    patch:
      tags:
        - Demo
      summary: Update demo active status
      operationId: updateDemoIsActive
      description: |
        Updates the demo is_active status.
        This endpoint allows updating only the is_active field.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - $ref: '#/components/parameters/WorkspaceIdHeader'
        - in: path
          name: demo_id
          required: true
          schema:
            type: string
            format: uuid
          description: Demo identifier
          example: "bd08612-1a9d-4a6e-8431-1demoid"
        
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                  description: Whether the demo is active
                  example: true
      responses:
        '200':
          description: Demo active status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Demo active status updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'


  


 
components:
  
  parameters:
    UserIdHeader:
      name: user_id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: ID of the authenticated user
      example: "c8d8869-4060-4e8a-ba30-c6da3buserid"
    
    WorkspaceIdHeader:
      name: workspace_id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: ID of the demo
      example: "bd08612-1a9d-4a6e-8431-1workspaceid"
  
  schemas:
    
    BaseAppModel:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: Timestamp when record was created
          example: "2024-01-15T10:30:00Z"
          x-read-only: true
          x-sortable: true
          x-filterable: true
          x-display-name: "Created At"
          x-ui-group: "Audit Information"
          x-ui-order: 101
          x-component-hint: "datetime-display"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when record was last updated
          example: "2024-01-15T10:30:00Z"
          x-read-only: true
          x-sortable: true
          x-filterable: true
          x-display-name: "Updated At"
          x-ui-group: "Audit Information"
          x-ui-order: 102
          x-component-hint: "datetime-display"
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when record was deleted
          example: "2024-01-15T10:30:00Z"
          x-read-only: true
          x-filterable: true
          x-display-name: "Deleted At"
          x-ui-group: "Audit Information"
          x-ui-order: 103
          x-component-hint: "datetime-display"
          x-hidden: true
        created_by:
          type: string
          format: uuid
          description: ID of user who created the record
          example: "c8d8869-4060-4e8a-ba30-c6da3buserid"
          x-read-only: true
          x-filterable: true
          x-display-name: "Created By"
          x-ui-group: "Audit Information"
          x-ui-order: 104
          x-component-hint: "user-selector"
          x-lookup-endpoint: "/api/v1/users/{value}"
        updated_by:
          type: string
          format: uuid
          nullable: true
          description: ID of user who last updated the record
          example: "c8d8869-4060-4e8a-ba30-c6da3buserid"
          x-read-only: true
          x-filterable: true
          x-display-name: "Updated By"
          x-ui-group: "Audit Information"
          x-ui-order: 105
          x-component-hint: "user-selector"
          x-lookup-endpoint: "/api/v1/users/{value}"
        deleted_by:
          type: string
          format: uuid
          nullable: true
          description: ID of user who deleted the record
          example: "c8d8869-4060-4e8a-ba30-c6da3buserid"
          x-read-only: true
          x-filterable: true
          x-display-name: "Deleted By"
          x-ui-group: "Audit Information"
          x-ui-order: 106
          x-component-hint: "user-selector"
          x-lookup-endpoint: "/api/v1/users/{value}"
          x-hidden: true
        status:
          type: string
          enum: ['creating', 'created', 'updating', 'updated', 'deleting', 'deleted']
          description: Current processing status of the record
          example: "created"
          x-sortable: true
          x-filterable: true
          x-display-name: "Status"
          x-ui-group: "Status Information"
          x-ui-order: 20
          x-component-hint: "select"
          x-state-machine:
            initial: "creating"
            states:
              creating:
                on:
                  SUCCESS: "created"
                  FAILURE: "deleted"
              created:
                on:
                  UPDATE: "updating"
                  DELETE: "deleting"
              updating:
                on:
                  SUCCESS: "updated"
                  FAILURE: "created"
        is_active:
          type: boolean
          description: Whether the record is active (not the same as status)
          example: true
          x-sortable: true
          x-filterable: true
          x-display-name: "Active"
          x-ui-group: "Status Information"
          x-ui-order: 21
          x-component-hint: "toggle-switch"
        error_message:
          type: string
          nullable: true
          description: Technical error message for debugging
          example: "Database connection timeout after 30 seconds"
          x-read-only: true
          x-display-name: "Error Message"
          x-ui-group: "Status Information"
          x-ui-order: 22
          x-component-hint: "textarea"
          x-hidden: true
          x-internal: true
          x-log-redact: true
        error_user_message:
          type: string
          nullable: true
          description: User-friendly error message for display
          example: "Something went wrong. Please try again later."
          x-read-only: true
          x-display-name: "User-Friendly Error"
          x-ui-group: "Status Information"
          x-ui-order: 23
          x-component-hint: "alert-message"
          x-hidden: true
      
    # Demo
    DemoModel:
      allOf:
        - $ref: '#/components/schemas/BaseAppModel'
        - type: object
          properties:
            demo_id:
              type: string
              format: uuid
              readOnly: true
              description: Demo identifier
              example: "bd08612-1a9d-4a6e-8431-1demoid"
              x-read-only: true
              x-filterable: true
              x-display-name: "Demo ID"
              x-ui-group: "Primary Details"
              x-ui-order: 1
              x-component-hint: "uuid-display"
            name:
              type: string
              maxLength: 200
              description: Demo name
              example: "Marketing Team"
              x-searchable: true
              x-sortable: true
              x-filterable: true
              x-display-name: "Demo Name"
              x-placeholder: "Enter the name of the demo"
              x-component-hint: "textfield"
              x-ui-group: "Primary Details"
              x-ui-order: 2
              x-validation-regex: "^[a-zA-Z0-9 ]{1,200}$"
              x-business-rule: "The demo name must be unique within the workspace."
          required:
            - demo_id

    
    # Pagination Meta
    PaginationMeta:
      type: object
      properties:
        total_count:
          type: integer
          example: 50
          description: Total number of records
        offset:
          type: integer
          example: 0
          description: Current offset
        limit:
          type: integer
          example: 100
          description: Number of items per request
        total_pages:
          type: integer
          example: 5
          description: Total number of pages
        
    

  # Responses
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error_message:
                type: string
                example: "Bad Request"

    NotFound:
      description: Resource Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error_message:
                type: string
                example: "Not Found"

    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error_message:
                type: string
                example: "Internal Server Error"
    
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error_message:
                type: string
                example: "Authentication credentials were not provided."

    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error_message:
                type: string
                example: "You do not have permission to perform this action."